+++ apps/web/src/app/(with-side-bar)/topic/[id]/sticker/page.tsx
@@ -4,10 +4,14 @@ interface TopicDetailProps {
   params: Promise<{
     id: string;
   }>;
+  searchParams: Promise<{
+    stickerId: string;
+  }>;
 }
 
-export default async function TopicStickerDetail({ params }: TopicDetailProps) {
+export default async function TopicStickerDetail({ params, searchParams }: TopicDetailProps) {
   const { id } = await params;
+  const { stickerId } = await searchParams;
 
-  return <TopicStickerDetailPage id={id} />;
+  return <TopicStickerDetailPage id={id} stickerId={stickerId} />;
 }
