+++ apps/web/src/components/(with-side-bar)/topic/remove-content-button.tsx
