+++ apps/web/src/components/(with-side-bar)/topic/remove-dialog-content.tsx
