+++ apps/web/src/components/(with-side-bar)/topic/sticker/block-note.tsx
