+++ apps/web/src/components/(with-side-bar)/topic/sticker/remove-user-sticker-dialog.tsx
