+++ apps/web/src/components/(with-side-bar)/topic/sticker/topic-sticker.tsx
