+++ apps/web/src/components/common/resize-bar.tsx
