+++ apps/web/src/page/topic/[id]/sticker/index.tsx
@@ -4,11 +4,16 @@ import dynamic from "next/dynamic";
 
 interface TopicStickerDetailPageProps {
   id: string;
+  stickerId: string;
 }
 
-const BlockNote = dynamic(() => import("@/components/block-note"), { ssr: false });
+const BlockNote = dynamic(() => import("@/components/(with-side-bar)/topic/sticker/block-note"), {
+  ssr: false,
+});
 
-export default function TopicStickerDetailPage({ id }: TopicStickerDetailPageProps) {
+export default function TopicStickerDetailPage({ id, stickerId }: TopicStickerDetailPageProps) {
+  // stickerId가 있다면 커스텀 스티커(노란색) 조회
+
   return (
     <main className="flex h-full flex-col gap-4">
       <input
